+++ app/src/main/java/com/aintogarov/currencyconverter/domain/MoneyAmountState.kt
@@ -1,7 +1,9 @@
 package com.aintogarov.currencyconverter.domain
 
+import java.math.BigDecimal
+
 
 data class MoneyAmountState(
     val currency: String,
-    val value: Double
+    val value: BigDecimal
 )
